backend/requirements.txt
# Web 框架
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.1
Flask-Migrate==4.0.7
Flask-SocketIO==5.3.6
eventlet==0.36.1
gunicorn==22.0.0

# 数据库驱动
PyMySQL==1.1.1
cryptography==42.0.8

# 环境变量
python-dotenv==1.0.1

# 密码加密
bcrypt==4.1.3

# Excel 处理
openpyxl==3.1.5

# AI 推理（核心）
torch==2.3.1
torchvision==0.18.1
facenet-pytorch==2.6.0
opencv-python-headless==4.10.0.84
Pillow==10.4.0
numpy==1.26.4

# 测试
pytest==8.2.2
pytest-cov==5.0.0
pytest-mock==3.14.0
